fixed tests

+++ tmcrstudioaddin/tests/testthat/testHTTPQueries.R
@@ -16,7 +16,9 @@ test_that("organizations are fetched from the server",{
     expect_true(is.character(organization))
   }
 
-  file.remove(".credentials.rds")
+  if(file.exists(".credentials.rds")){
+    file.remove(".credentials.rds")
+  }
 })
 test_that("empty organization list is returned if failed",{
   if(file.exists(".credentials.rds")){
@@ -26,7 +28,9 @@ test_that("empty organization list is returned if failed",{
   organizations<-getAllOrganizations()
   expect_true(length(organizations)==0)
 
-  file.remove(".credentials.rds")
+  if(file.exists(".credentials.rds")){
+    file.remove(".credentials.rds")
+  }
 })
 test_that("courses associated with server and organization are fetched from the server",{
   if(file.exists(".credentials.rds")){
@@ -39,7 +43,9 @@ test_that("courses associated with server and organization are fetched from the 
     expect_true(is.character(course))
   }
 
-  file.remove(".credentials.rds")
+  if(file.exists(".credentials.rds")){
+    file.remove(".credentials.rds")
+  }
 })
 test_that("empty courses list is returned if failure",{
   if(file.exists(".credentials.rds")){
@@ -49,5 +55,7 @@ test_that("empty courses list is returned if failure",{
   courses<-getAllCourses("hy")
   expect_true(length(courses)==0)
 
-  file.remove(".credentials.rds")
+  if(file.exists(".credentials.rds")){
+    file.remove(".credentials.rds")
+  }
 })
